requirements.txt
streamlit
pandas
pydantic
python-dotenv
rapidfuzz
pypdf
boto3